Former-commit-id: 104f4cb389068eed3fb1486833a4f1a678fe648f

+++ id/src/main/java/com/simprints/id/data/DataManager.kt
@@ -1,13 +1,13 @@
 package com.simprints.id.data
 
 import com.simprints.id.data.analytics.AnalyticsManager
+import com.simprints.id.data.db.DataCallback
 import com.simprints.id.data.db.DbManager
+import com.simprints.id.data.db.remote.enums.VERIFY_GUID_EXISTS_RESULT
 import com.simprints.id.data.prefs.PreferencesManager
 import com.simprints.id.data.secure.SecureDataManager
-import com.simprints.id.data.db.DataCallback
-import com.simprints.id.data.db.remote.enums.VERIFY_GUID_EXISTS_RESULT
-import com.simprints.id.domain.Constants
 import com.simprints.id.domain.ALERT_TYPE
+import com.simprints.id.domain.Constants
 import com.simprints.libcommon.Person
 import com.simprints.libsimprints.Identification
 import com.simprints.libsimprints.RefusalForm
@@ -38,7 +38,7 @@ interface DataManager : PreferencesManager, DbManager,
     // Data transfer
     fun savePerson(person: Person)
     fun loadPeople(destinationList: MutableList<Person>, group: Constants.GROUP, callback: DataCallback?)
-    fun getPeopleCount(group: Constants.GROUP): Long
+    fun getPeopleCount(group: Constants.GROUP): Int
 
     fun saveIdentification(probe: Person, matchSize: Int, matches: List<Identification>)
     fun updateIdentification(projectId: String, selectedGuid: String)
