+++ id/src/main/java/com/simprints/id/data/DataManagerImpl.kt
@@ -108,11 +108,11 @@ class DataManagerImpl(private val context: Context,
         dbManager.loadPeople(destinationList, group, userId, moduleId, callback)
     }
 
-    override fun getPeopleCount(group: Constants.GROUP): Long =
+    override fun getPeopleCount(group: Constants.GROUP): Int =
         when (group) {
             Constants.GROUP.GLOBAL -> dbManager.getPeopleCount()
             Constants.GROUP.USER -> dbManager.getPeopleCount(userId = userId)
-            Constants.GROUP.MODULE ->dbManager.getPeopleCount(userId = userId, moduleId = moduleId)
+            Constants.GROUP.MODULE -> dbManager.getPeopleCount(userId = userId, moduleId = moduleId)
         }
 
     override fun saveIdentification(probe: Person, matchSize: Int, matches: List<Identification>) {
